FGVC-Aircraft_manufacturers/requirements.txt
gradio
torch
torchvision
fastai
pillow
numpy
ipython